Move ArchiveStreamReader and its large dependency from vespajlib to application-model. This avoids it being pulled in to the config-model-fat that make the install larger than necessary.
