+++ application-model/src/main/java/com/yahoo/vespa/archive/ArchiveStreamReader.java
@@ -1,5 +1,5 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.compress;
+package com.yahoo.vespa.archive;
 
 import com.yahoo.path.Path;
 import com.yahoo.yolean.Exceptions;
