+++ cloud-tenant-base-dependencies-enforcer/pom.xml
@@ -186,7 +186,6 @@
                                         <include>org.antlr:antlr-runtime:3.5.2:test</include>
                                         <include>org.antlr:antlr4-runtime:4.9.3:test</include>
                                         <include>org.apache.commons:commons-exec:1.3:test</include>
-                                        <include>org.apache.commons:commons-compress:1.21:test</include>
                                         <include>org.apache.commons:commons-math3:3.6.1:test</include>
                                         <include>org.apache.felix:org.apache.felix.framework:${felix.version}:test</include>
                                         <include>org.apache.felix:org.apache.felix.log:1.0.1:test</include>
